// src/collection_card/dto/raw-collection_card.dto.ts
export class RawCollectionCardDto {
  collection_card_id: number;
  collection_card_tradeable: number;
  collection_card_foil: number;
  collection_card_collectionId: number;
  collection_card_cardId: number;
  card_id: number;
  card_scryfallId: string;
  card_name: string;
  card_cardType: string;
  card_setName: string;
  card_setCode: string;
  card_flatValue: string;
  card_foilValue: string;
  card_color: string;
  card_colorIdentity: string;
  card_cmc: number;
  card_rarity: string;
  card_cardUri: string;
  card_artUri: string;
  card_finishes: string;
  card_language: string;
  card_collectorNumber: string;
  collection_id: number;
  collection_public: number;
  collection_userId: number;
  value: string;
}